+++ src/components/layout/footer/Footer.jsx
@@ -1,9 +1,9 @@
-import '../../styles/Footer.css';
-import IconWithLink from './footer/IconWithLink';
-import facebookIcon from '../../assets/facebook.svg';
-import twitterIcon from '../../assets/twitter.svg';
-import youtubeIcon from '../../assets/youtube.svg';
-import instagramIcon from '../../assets/instagram.svg';
+import './Footer.css';
+import IconWithLink from './IconWithLink';
+import facebookIcon from '../../../assets/facebook.svg';
+import twitterIcon from '../../../assets/twitter.svg';
+import youtubeIcon from '../../../assets/youtube.svg';
+import instagramIcon from '../../../assets/instagram.svg';
 
 const SNS_ITEMS = [
   {
