+++ src/components/layout/header/Header.css
@@ -35,6 +35,17 @@
   gap: 6px;
 }
 
+.btn {
+  height: 5.4rem;
+  border-radius: 0.8rem;
+  background: linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%);
+  line-height: 5.4rem;
+  font-size: 1.8rem;
+  font-weight: 600;
+  color: var(--color-gray-100);
+  text-align: center;
+}
+
 .user-profile {
   width: 2.8rem;
   height: 2.8rem;
