--- src/styles/FolderSection.css
@@ -1,93 +1,0 @@
-.section-area {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: 4rem 0;
-}
-
-.search-bar-group {
-  display: flex;
-  align-items: center;
-  gap: 1rem;
-  margin-bottom: 4rem;
-  padding: 1.5rem 1.6rem;
-  width: 100%;
-  max-width: 106rem;
-  border-radius: 1rem;
-  background-color: var(--color-gray-100);
-}
-
-.icon-search {
-  width: 1.6rem;
-  height: 1.6rem;
-}
-
-.input-search {
-  width: 100%;
-  background-color: inherit;
-  line-height: 2.4rem;
-  font-size: 1.6rem;
-  font-weight: 400;
-  outline: none;
-}
-
-.card-list {
-  display: flex;
-  flex-wrap: wrap;
-  margin: 0 auto;
-  max-width: 106rem;
-  gap: 2.5rem 2rem;
-}
-
-.card-item {
-  width: 34rem;
-  border-radius: 1.5rem;
-  box-shadow: 0 0.5rem 2.5rem 0 rgba(0, 0, 0, 0.08);
-  overflow: hidden;
-}
-
-.card-item:hover .card-info-group {
-  background-color: #f0f6ff;
-}
-
-.card-item:hover .img-card {
-  transform: scale(1.3);
-}
-
-@media (max-width: 1199px) {
-  .section-area {
-    padding: 4rem 3.2rem;
-  }
-
-  .search-bar-group {
-    max-width: 70.4rem;
-  }
-
-  .card-list {
-    width: 70.4rem;
-    gap: 2.5rem 2.4rem;
-  }
-}
-
-@media (max-width: 767px) {
-  .search-bar-group {
-    gap: 0.6rem;
-    margin-bottom: 3.2rem;
-    padding: 1.3rem 1.6rem;
-  }
-
-  .input-search {
-    width: 100%;
-    line-height: 1.7rem;
-    font-size: 1.4rem;
-  }
-
-  .card-list {
-    width: 100%;
-    gap: 2rem;
-  }
-
-  .card-item {
-    width: 100%;
-  }
-}
