--- src/styles/common.css
@@ -1,10 +1,0 @@
-.btn {
-  height: 5.4rem;
-  border-radius: 0.8rem;
-  background: var(--gra-purpleblue-to-skyblue, linear-gradient(91deg, #6d6afe 0.12%, #6ae3fe 101.84%));
-  line-height: 5.4rem;
-  font-size: 1.8rem;
-  font-weight: 600;
-  color: var(--color-gray-100);
-  text-align: center;
-}
