+++ src/styles/style.css
@@ -1,3 +1,2 @@
 @import url('./variables.css');
 @import url('./reset.css');
-@import url('./common.css');
